requirements.txt
# 🔹 Flask (API Framework)
Flask==3.0.0

# 🔹 Google Cloud Firestore (Database)
google-cloud-firestore==2.11.0

# 🔹 Twilio (SMS Integration)
twilio==9.0.1

# 🔹 Dialogflow CX (AI Chatbot Integration)
google-cloud-dialogflow-cx==1.12.0

# 🔹 Salesforce (Customer Relationship Management)
simple-salesforce==1.12.3

# 🔹 Environment Variables Management
python-dotenv==1.0.0

# 🔹 Testing & Mocking
pytest==8.0.0
pytest-mock==3.12.0
unittest2==1.1.0

# 🔹 HTTP Requests (for API calls)
requests==2.31.0